287 Move search indexing into separate class/method
